Android版 BeaconService最終調整 BeaconPageViewModelとの接続完了、検出、送出OK
