+++ Covid19Radar/Covid19Radar.Android/Services/RangeEventArgs.cs
@@ -1,0 +1,15 @@
+﻿using AltBeaconOrg.BoundBeacon;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+
+namespace Covid19Radar.Droid.Services
+{
+    public class RangeEventArgs : EventArgs
+    {
+        public Region Region { get; set; }
+        public ICollection<Beacon> Beacons { get; set; }
+    }
+}
